pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rbi-nbfc-chatbot"
version = "0.1.0"
description = "RAG-powered chatbot for RBI NBFC regulations"
requires-python = ">=3.9"
dependencies = [
    "langchain",
    "langchain-community",
    "langchain-google-genai",
    "faiss-cpu",
    "streamlit",
    "fastapi",
    "uvicorn",
    "python-dotenv",
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
